requirements.txt
openai >= 0.27.2
google-cloud-texttospeech >= 2.14.1
pydub >= 0.25.1
pyaudio >= 0.2.12
sounddevice >= 0.4.6
scipy >= 1.10.1
audioread >= 3.0.0